backend/virtual-tryon/requirements.txt
boto3==1.34.0
Pillow==10.2.0
